Add the `adder` tool gem.
This tool will be the example for the first exposure to the BDD / TDD
process. It is setup with some basic structure to facilitate the process
without requiring lots of extra custom configurations.

RSpec is being used for the testing tool. All TDD / BDD can be done with
most modern testing tools. BDD does not require a tool such as Cucumber,
but Cucumber, and other similar tools, can help facility communication
with end users.

For this tool, all testing will be done in RSpec. A sample `.rspec` file
is included to show how some custom settings can be added. Additionally,
a `spec_helper` is updated to allow for some BDD type specific
enhancements; such as filtering out any works in progress.