Start with a high level feature specification.
With BDD, focus is generally on a top-down approach. It also focuses on
how end-users interact with the program. It is not concerned necessarily
with implementation details.

Using this approach, the feature specs are based around user / job
stories. They view the program as more of a black box. Due to this, most
tests tend to be full integration tests where no mocks or stubs are
used. Instead, the focus is solely on environment state changes.

Since some features can a involved and require lots of new code, they
are generally marked as `wip` for "work in progress". This allows the
incomplete feature to be excluded from the main test run, while it is
implemented.