Add executable for the cli.
Here we add the `adder` executable for the CLI. We setup the shebang to
run the available version of Ruby. We also set the executable bit on it.

Print the usage when the executable is run. We can run this locally by
executing "bin/adder". The `spec_helper` already adjusts our path for us
so that we can omit the `bin/` path in tests. Once the gem is installed
we can let bundler handle adding the `bin/` path for us.

Lastly, the wip tag is removed on the spec as it is fully passing.