Add missing class and methods.
We add just enough code to make the feature spec finish. It is still
failing, but it is no longer raising exceptions related to missing
classes and methods.

At this point, the feature spec is waiting for the lower level classes
to be properly implement. It will need to stay as `wip` until we finish
this.

We need to now drop into more unit level tests in order to continue the
program.