Add failing sum unit test.
Once again it's time to write the failing test first. This type of test
is a unit test. It's only testing a single object. The goals with unit
tests is, given various states, verify behavior of a single object.

We've also, once again, marked the test as a `wip` as it's failing; and
we don't want to break the build. In a normal system, this commit will
most likely not exist. Instead it will be committed along with the
passing behavior having been implemented. It is being committed
standalone to clearly delineate that we are writing the failing unit
test first.

Future commits may merge the behaviors together and only note in the
commit that the test was written first.