Implement the adding machine sum behavior.
Many people do not write comments in Ruby, myself included. However, in
this case, I felt it necessary to help spell out the full spec of the
method as we defined it (in our heads and explanatory comments) in the
executable.

While the test only passes in a single number, we've written the method
definition to support multiple numbers. It would be more work to just
strip off the first number and adjust it, than it was to write the
implementation that we did. Again, "strict TDD" would dictate that we
write something like:

   BigDecimal.new(numbers.first)

As with most things, there are shades of gray to the processes. In my
experience, to date, there are little added benefits from these tiny
steps for simple methods. For more complex ones, or others where the
algorithm is not readily apparent, I very much follow the stricter TDD
style to help flush out the algorithm that is needed.

Since the test is now passing we've removed the `wip` tag.