Flush out remaining parts of summing algorithm.
You may have noticed that our integration test is now passing. However,
it's still marked as `wip`. The reason for this is, I'm avoiding the
full BDD cycle. For me, I have the full spec of the
`Adder::Machine#sum` already. I dislike popping up the BDD stack when I
haven't fully implemented the known spec for the behavior under test.

As we go through and test things, I want those tests to reflect my
knowledge of the system at that time. If I miss a test case, it's
usually because I forgot about it. To me popping up the stack to the
feature specs at this point leaves the `Adder::Machine#sum` in a state
that is not fully reflective of our knowledge. We know the full spec, we
should finish it now (in the real world, I've seen methods left
incomplete as someone popped up the stack and never came back to finish
it as they were pulled onto something else).

As we've been doing, we write our test first. Instead of writing many
different test cases for each of our contexts (one for each of the
different things we could pass in: Fixnum, Float, String), we cheat. We
carefully pick a test case that merges all of these behaviors together.
Now there are some trade-offs we're making here:

  * When the test fails, it may be harder to pinpoint exactly which
    argument is messing with things
  * We don't have to duplicate the test cases for each context
  * We are trusting that this level of coverage is enough

In normal day to day activities, these types of trade-offs are made
constantly. If there is ever a bug encountered, go back and write a test
for it!

Here we find out that our implementation isn't actually correct. The
first error we encounter is one of passing an `Array` to `BigDecimal`.
This is the same error that our feature spec is now throwing at us. The
fix is simple enough: use `flatten` to make the `numbers` array one
dimensional. Note though that `flatten` is notoriously slow. We'll
accept this trade off for now as we never expect the `numbers` array to
be large. Nor is squeezing every once of speed out of this a priority.

Next we encounter an error that we've omitted precision for the float
when provided to `BigDecimal`. Since we want to support whatever
precision that the client gives us we use a little hack. By converting
the numbers into `String`s we leave it to `BigDecimal` to do the
conversions for us.

After these changes, everything is passing and we can continue.