Finish adder sum specs.
Based on our current understanding of the `Adder::Machine#sum` spec
there is one last case to test: passing a bad value and having it become
a zero. Our current implementation already handles this case for us!
(well technically `BigDecimal::new` does).

Always be suspicious of tests that pass the first time around. I like to
change the code under test to make the test fail to verify I'm not
losing my mind and that I have the proper understanding of what
mechanism is causing stuff to pass. However, in this case, the act of
passing the `String` to `BigDecimal` does this. There isn't much of a
code change we can make without breaking it all.

In this case, I open a `pry` session and manually verify that
`BigDecimal` is what is doing the conversion. That knowledge, combined
with the passing test, is enough _trust_ for me to continue.

Now that our spec is fully implemented, we pop up the BDD stack to the
feature spec. It's now passing too! So we remove the `wip` tag.