Add remaining feature specs.
Here we round out the rest of our desired features per the README.
Again, as we write each of these specs, one by one, we test them. Since
we've done the implementation they should pass, unless we made a
mistake. To verify this, we temporarily comment out or adjust the line
of code that should be making the test pass (the `.round(2)`).

Since, the tests are passing there is no need to mark them with `wip`.
Also, since our tests are straight forward and involve no real code
changes, we can combine them into a single happy commit.

We're now done with our adder command line tool.

:sweat_smile: :thumbsup: