Add the `trader` tool gem.
This tool will be our 2nd example using the BDD / TDD process.  It is setup
with some basic structure to facilitate the process without requiring lots of
extra custom configurations.

In this example, both Cucumber and RSpec will be used. We choose to use Cucumber
here to demonstrate it as a tool, and to show how the `aruba` gem makes it easy
to write public specifications for the tool.

Some basic setup for both Cucumber and RSpec have been included. See the
corresponding `cucumber.yml`, `.rspec-example`, and `spec/spec_helper.rb` for
details.