Start with a high level feature specification.
Again, we start with some general BDD. We add a high level spec to get all of
our parts wired up. The goal here is to make sure we have Cucumber working and
the executable tool setup.

Aruba is kind enough to provide us with all of the Cucumber steps that we will
need for the majority of our CLI tests.

Here we've added a very high level description of how the tool works (similar
to what's in the README). We then add the initial scenario of running the
command. Aruba has already added our `bin` directory to it's load path, so it
will look there to try to run it.

As with all feature specs that we are working we tag it with a `wip` flag.
Cucumber expects this to precede the scenario and begin with an '@'.