Add executable for the cli.
Here we add the `trader` executable for the CLI. We setup the shebang to
run the available version of Ruby. We also set the executable bit on it.

Print the usage when the executable is run. We can run this locally by
executing "bin/trader". Aruba has already adjusted our path for us, so it sees
the file when it runs the spec.

Lastly, the wip tag is removed on the spec as it is fully passing.