Add the next general usage job story.
For the scenario we provide some details via free form text on what the
job story actually is. We then write the scenario using the Aruba steps
to create files with specific content. The our executable is run and the
output is compared.

Once again, as this is a feature spec, we mark it with `wip` until the
feature is complete.