Write the desired implementation of the feature.
This is our chance to write the code we wish we had. It doesn't matter
if none of it exists. We will create it. This is our chance to define
our internal API for what objects we have, how we interface with them,
and what they return to us.

Our existing spec still passes; though the wip one fails. This is fine.
We expect the wip specs to fail on some commits, that's why we marked
them. Since our existing spec is passing we can safely commit this
partial feature implementation with out worrying that the existing app
is broken. We'll continue to follow this process as we build out the
features.