Add the exchange shell.
Once again, we're not really sure how we want this object to work. We
aren't directly interacting with it yet, just passing things around.

Our tests reflect our lack of knowledge, though this is the current
understanding of the system. There is nothing wrong with this. It shows
that we aren't sure what is going to happen yet.